# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CampusRide API
  description: |
    CampusRide backend: REST API for inter-campus ride sharing.

    Response envelope (all endpoints):
    - success: `{ "success": true, "data": {...}, "meta": null }`
    - error: `{ "success": false, "error": { "code": "...", "message": "...", "fields": [...] } }`

    Common error codes: `VALIDATION_ERROR`, `NOT_FOUND`, `UNAUTHORIZED`, `FORBIDDEN`,
    `EMAIL_TAKEN`, `STUDENT_ID_TAKEN`, `AUTH_INVALID_CREDENTIALS`, `AUTH_TOKEN_INVALID`,
    `AUTH_REFRESH_REVOKED`, `ACCOUNT_SUSPENDED`, `DRIVER_NOT_VERIFIED`,
    `VEHICLE_NOT_VERIFIED`, `CAMPUS_NOT_FOUND`, `RIDE_NOT_FOUND`, `RIDE_FULL`,
    `RIDE_INVALID_STATE`, `BOOKING_NOT_FOUND`, `BOOKING_ALREADY_EXISTS`,
    `BOOKING_INVALID_STATE`, `INVALID_PIN`, `LOCATION_NOT_ALLOWED`, `RATE_LIMITED`,
    `INTERNAL_ERROR`.

    Authenticate with `Authorization: Bearer <accessToken>`. Tokens are issued by
    `POST /auth/login` / `POST /auth/register` and rotated by `POST /auth/refresh`.
  version: 0.1.0
  contact:
    name: CampusRide Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /rides/{rideId}/bookings:
    post:
      summary: Request a booking on an open ride
      operationId: createBooking
      tags: [Bookings]
      parameters:
        - $ref: "#/components/parameters/RideIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [seats]
              properties:
                seats:
                  type: integer
                  minimum: 1
                  maximum: 20
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
    get:
      summary: List bookings for a ride (driver)
      operationId: listRideBookings
      tags: [Bookings]
      parameters:
        - $ref: "#/components/parameters/RideIdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
...
  /bookings/{id}:
    get:
      summary: Get a booking (participants only)
      operationId: getBooking
      tags: [Bookings]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /bookings/{id}/accept:
    post:
      summary: Accept a booking request (driver only; decrements seats atomically)
      operationId: acceptBooking
      tags: [Bookings]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
  /bookings/{id}/reject:
    post:
      summary: Reject a booking request (driver only)
      operationId: rejectBooking
      tags: [Bookings]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /bookings/{id}/cancel:
    post:
      summary: Cancel a booking (passenger or driver; restores seats)
      operationId: cancelBooking
      tags: [Bookings]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /bookings/{id}/checkin:
    post:
      summary: Check a passenger in with the ride PIN (driver only)
      operationId: checkinBooking
      tags: [Bookings]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [checkinPin]
              properties:
                checkinPin:
                  type: string
                  description: 4-digit ride PIN shown on the ride detail
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/Error"
  /bookings/{id}/no-show:
    post:
      summary: Mark a booking as no-show (driver only)
      operationId: noShowBooking
      tags: [Bookings]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /ride-requests:
    post:
      summary: Publish an open ride request (student)
      operationId: createRideRequest
      tags: [RideRequests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RideRequestPayload"
      responses:
        "201":
          $ref: "#/components/responses/Success"
...
  /ride-requests/{id}:
    get:
      summary: Get a ride request
      operationId: getRideRequest
      tags: [RideRequests]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
    patch:
      summary: Update an open ride request
      operationId: updateRideRequest
      tags: [RideRequests]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRideRequestPayload"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /ride-requests/{id}/cancel:
    post:
      summary: Cancel a ride request
      operationId: cancelRideRequest
      tags: [RideRequests]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /ride-requests/{id}/matches:
    get:
      summary: Matching open rides for a request
      operationId: rideRequestMatches
      tags: [RideRequests]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /rides/{rideId}/ratings:
    post:
      summary: Rate a participant of a completed ride
      operationId: createRating
      tags: [Ratings]
      parameters:
        - $ref: "#/components/parameters/RideIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [revieweeId, rating]
              properties:
                revieweeId:
                  type: string
                  format: uuid
                rating:
                  type: number
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
                  maxLength: 500
      responses:
        "201":
          $ref: "#/components/responses/Success"
  /users/{userId}/ratings:
    get:
      summary: Ratings received by a user
      operationId: ratingsForUser
      tags: [Ratings]
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
...
  /notifications:
    get:
      summary: List the user's notifications (newest first)
      operationId: listNotifications
      tags: [Notifications]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
...
  /notifications/{id}/read:
    patch:
      summary: Mark one notification as read
      operationId: markNotificationRead
      tags: [Notifications]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
...
  /admin/users/{userId}/status:
    patch:
      summary: Set account status (suspend/restore)
      operationId: setUserStatus
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accountStatus]
              properties:
                accountStatus:
                  type: string
                  enum: [ACTIVE, SUSPENDED]
      responses:
        "200":
          $ref: "#/components/responses/Success"
...
  /admin/drivers/{userId}/verification:
    patch:
      summary: Verify or reject a driver
      operationId: verifyDriver
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [VERIFIED, REJECTED]
      responses:
        "200":
          $ref: "#/components/responses/Success"
...
  /admin/vehicles/{vehicleId}/verification:
    patch:
      summary: Verify or reject a vehicle
      operationId: verifyVehicle
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/VehicleIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [VERIFIED, REJECTED]
      responses:
        "200":
          $ref: "#/components/responses/Success"
...
  /admin/reports/{id}:
    patch:
      summary: Resolve or dismiss a report
      operationId: resolveReport
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [RESOLVED, DISMISSED]
                adminAction:
                  type: string
                  maxLength: 1000
      responses:
        "200":
          $ref: "#/components/responses/Success"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RideIdParam:
      name: rideId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserIdParam:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VehicleIdParam:
      name: vehicleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    Success:
      description: Successful response with the standard envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiSuccess"
    Error:
      description: Failed response with the standard envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiSuccess:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          const: true
        data:
          nullable: true
        meta:
          nullable: true
    ApiError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message, fields]
          properties:
            code:
              type: string
              example: RIDE_FULL
            message:
              type: string
            fields:
              type: array
              items:
                type: object
                properties:
                  path:
                    type: string
                  message:
                    type: string
              nullable: true
...
    RideRequestPayload:
      type: object
      required: [originCampusId, destinationCampusId, earliestDeparture, latestDeparture, requiredSeats]
      properties:
        originCampusId:
          type: string
          format: uuid
        destinationCampusId:
          type: string
          format: uuid
        earliestDeparture:
          type: string
          format: date-time
        latestDeparture:
          type: string
          format: date-time
        requiredArrival:
          type: string
          format: date-time
        requiredSeats:
          type: integer
          minimum: 1
          maximum: 20
        maxBudget:
          type: number
          minimum: 1
        maxPickupDistanceKm:
          type: number
          minimum: 0
          maximum: 50
          default: 3
    UpdateRideRequestPayload:
      type: object
      properties:
        earliestDeparture:
          type: string
          format: date-time
        latestDeparture:
          type: string
          format: date-time
        requiredArrival:
          type: string
          format: date-time
          nullable: true
        requiredSeats:
          type: integer
          minimum: 1
          maximum: 20
        maxBudget:
          type: number
          minimum: 1
        maxPickupDistanceKm:
          type: number
          minimum: 0
          maximum: 50